+++ app/models/surveys/answer_template.rb
@@ -1,13 +1,19 @@
 class AnswerTemplate < ActiveRecord::Base
   # Constants
-  DATA_TYPE = ["answer_option_id", "numeric_value", "text_value"]
+  DATA_TYPES = ["answer_option_id", "numeric_value", "text_value"]
+  TEMPLATE_NAMES = ["date", "radio", "checkbox", "string", "height", "number"]
 
   # Concerns
   include Deletable
 
+  # Callbacks
+  before_validation :set_data_type
+  before_validation :set_allow_multiple
+
   # Model Validation
-  validates_presence_of :name, :data_type, :user_id
+  validates_presence_of :name, :data_type, :user_id, :template_name
   validates_uniqueness_of :name, scope: [ :deleted ]
+  validates_inclusion_of :template_name, in: TEMPLATE_NAMES
 
   # Model Relationships
   belongs_to :user
@@ -21,15 +27,20 @@ class AnswerTemplate < ActiveRecord::Base
   # Model Methods
 
   # Temporary Function
-  def template_name
+  def set_template_name!
+    new_template_name = nil
     question = self.questions.first
     if question
-      if question.display_type == "custom_date_input"        and self.data_type == "text_value"
+      new_template_name = if question.display_type == "custom_date_input" and self.data_type == "text_value"
         'date'
       elsif question.display_type == "radio_input"           and self.data_type == "answer_option_id"
         'radio'
       elsif question.display_type == "checkbox_input"        and self.data_type == "answer_option_id"
-        'checkbox'
+        if self.target_answer_option == nil
+          'checkbox'
+        else
+          'radio'
+        end
       elsif question.display_type == "checkbox_input"        and self.data_type == "text_value"
         'string'
       elsif question.display_type == "height_input"          and self.data_type == "numeric_value"
@@ -46,6 +57,29 @@ class AnswerTemplate < ActiveRecord::Base
     else
       nil
     end
+    self.update template_name: new_template_name
+  end
+
+  # Will replace "data_type" and be renamed to simply "data_type" when data_type
+  # is removed as a database column.
+  def stored_data_type
+    case self.template_name when 'number', 'height'
+      'numeric_value'
+    when 'radio', 'checkbox'
+      'answer_option_id'
+    else # 'date', 'string'
+      'text_value'
+    end
+  end
+
+  def set_data_type
+    self.data_type = self.stored_data_type
+    true # Before Validations need to return true in order to save record
+  end
+
+  def set_allow_multiple
+    self.allow_multiple = (self.template_name == 'checkbox')
+    true # Before Validations need to return true in order to save record
   end
 
   def target_templates(question, value)