+++ db/migrate/20150805150811_add_template_name_to_answer_templates.rb
@@ -1,0 +1,5 @@
+class AddTemplateNameToAnswerTemplates < ActiveRecord::Migration
+  def change
+    add_column :answer_templates, :template_name, :string
+  end
+end